seperated out the functions to different files. with archibus_functions.py just being the traffic cop
