+++ app/api/.env.example
@@ -14,3 +14,10 @@ ALLOWED_TOOLS=geds,corporate,archibus
 AZURE_SEARCH_INDEX_NAME=current
 ARCHIBUS_API_USERNAME=user
 ARCHIBUS_API_PASSWORD=pass
+
+#Archibus vendor API credentials - DEV
+ARCHIBUS_API_V1_TENANT=something
+ARCHIBUS_API_V1_CLIENT=something
+ARCHIBUS_API_V1_CLIENT_SECRET=something
+ARCHIBUS_API_V1_APPLICATION_ID=something
+ARCHBIUS_API_V1_URL=http://localhost:8081/archibus/api/
